% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\name{data_description}
\alias{data_description}
\title{data_description}
\usage{
data_description(data, variable = NULL, description = NULL)
}
\arguments{
\item{data}{A dataframe whose variables are to be described.}

\item{variable}{A vector of variable names from the dataframe. If provided, descriptions will be updated for these variables.}

\item{description}{A vector of descriptions. If `variable` is not provided, descriptions will be applied sequentially to the first N columns.
If `variable` is provided, descriptions will be applied to the corresponding variables.}
}
\value{
A dataframe with the variables and their descriptions.
}
\description{
Create a data description dataframe and optionally update the descriptions
}
\details{
#' This function generates a dataframe of variables from the dataset and their descriptions.
If `description` is provided without `variable`, it updates the descriptions sequentially for the first N columns.
If both `variable` and `description` are provided, it updates the specified variables with the corresponding descriptions.
}
\examples{
# Create a description dataframe without updating any descriptions
descriptions <- data_description(iris)
print(descriptions)

# Update descriptions for the first two columns
descriptions <- data_description(iris, description = c("Length of sepals", "Width of sepals"))
print(descriptions)

# Update descriptions for specific variables
descriptions <- data_description(iris,
                                 variable = c("Sepal.Length", "Species"),
                                 description = c("Length of sepals", "Species of the flower"))
print(descriptions)
}
