% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\name{data_dictionary}
\alias{data_dictionary}
\title{#' data_dictionary}
\usage{
data_dictionary(
  data,
  data_title = NULL,
  descriptions = NULL,
  show_na = TRUE,
  top_n = 5
)
}
\arguments{
\item{data}{A dataframe for which the data dictionary will be printed.}

\item{data_title}{A title for the data dictionary.}

\item{descriptions}{A dataframe containing descriptions of the variables. It should have at least two columns: "Variable" and "Description".}

\item{show_na}{Logical; whether to display NA counts and percentages in the data dictionary. Default is `TRUE`.}

\item{top_n}{Integer; the number of top unique values to display for each column. Default is 5.}
}
\value{
None. Prints the data dictionary.
}
\description{
Prints a data dictionary for a given dataset in markdown format.
}
\details{
This function generates a data dictionary for the provided dataset. Optionally, it includes column descriptions, counts of missing values, and the top unique values for each column.
}
\examples{
# Create a description dataframe for iris and update one of the descriptions
desc <- data_description(iris, variable = c("Species"), description = c("Species of Flower"))

# Print the data dictionary for the iris dataset
data_dictionary(iris, descriptions = desc, top_n = 1)
}
\seealso{
\code{\link[explore]{data_dict_md}} for the original function in the `explore` package.
}
