% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{data_summary}
\alias{data_summary}
\title{data_summary}
\usage{
data_summary(data, var_types = NULL, include_stats = NULL, data_title = NULL)
}
\arguments{
\item{data}{A dataframe containing the variables for which summary statistics will be generated.}

\item{var_types}{A named list of variable types. If not provided, all types will be used.}

\item{include_stats}{A list of statistics to include for each variable type (e.g., mean, median, sd). If not provided, default summary statistics will be computed based on the variable type.}

\item{data_title}{A title for the dataset. If not provided it will be empty.}
}
\value{
A dataframe or printed output summarizing the statistics for each variable, depending on the variable type.
}
\description{
Generate Summary for Different Variable Types
}
\details{
This function generates summary statistics for variables in a dataset, categorized by their types. It allows for custom selection of variable types and summary statistics to include. The function can handle numeric, character, and other variable types and compute appropriate summary statistics for each.
}
\examples{
# Example usage with the iris dataset
summary_stats <- data_summary(iris)
print(summary_stats)

# Specify variable types manually and include custom statistics
var_types <- list(Sepal.Length = "numeric", Species = "character")
include_stats <- list(numeric = c("mean", "median", "sd"), character = c("count"))
summary_stats <- data_summary(iris, var_types = var_types, include_stats = include_stats, data_title = "Iris Summary")
print(summary_stats)

}
