% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataDict.R
\name{update_description}
\alias{update_description}
\title{Update Descriptions for Variables}
\usage{
update_description(descriptions, variable, description)
}
\arguments{
\item{descriptions}{A dataframe containing variable descriptions.}

\item{variable}{The name of the variable for which the description is to be updated.}

\item{description}{The new description for the variable.}
}
\value{
The updated descriptions dataframe.
}
\description{
Update Descriptions for Variables
}
\examples{
descriptions <- update_description(descriptions,
                                   "osm_id",
                                   "Unique Identifier from OpenStreetMap")
descriptions <- update_description(descriptions,
                                   c("osm_id", "name"),
                                   c("Unique Identifier from OpenStreetMap",
                                     "Name of the entity"))
}
