% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataDict.R
\name{print_data_dict}
\alias{print_data_dict}
\title{Print Data Dictionary}
\usage{
print_data_dict(data, title = "", descriptions = NULL)
}
\arguments{
\item{data}{The dataset for which the data dictionary is to be printed.}

\item{title}{The title for the data dictionary (optional).}

\item{descriptions}{A data frame providing detailed descriptions for
variables (optional).}
}
\value{
None
}
\description{
Creates a data dictionary for the dataset using attributes from the data
and description if provided.
Adapted from the data_dict_md function from the explore package.
Source: https://rdrr.io/cran/explore/src/R/tools.R
}
\examples{
descriptions = print_descriptions_df(mtcars)
print_data_dict(mtcars, title = "Cars", descriptions = descriptions)
}
